+++ .env.example
@@ -1,4 +1,5 @@
 SESSION_COOKIE_SECRET=some-secret
+JIRA_TICKET_BASE_URL=https://jira.atlassian.net/browse
 DATABASE_URL="mysql://root:password@localhost:3306/code_review_bot"
 SLACK_BOT_TOKEN=xoxb-your-bot-token
 SLACK_SIGNING_SECRET=your-signing-secret
